@@ -1,6 +1,6 @@
 package com.tozny.androide3db
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 
 class DecryptPhotoActivity : AppCompatActivity() {
