@@ -1,7 +1,7 @@
 package com.tozny.androide3db
 
 import android.app.ActionBar
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
