@@ -1,7 +1,7 @@
 package com.tozny.androide3db
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.text.Editable
 import android.util.Log
 import android.view.View
